Precompute rust types, and reduce heap allocations
