+++ gl_generator/registry/mod.rs
@@ -14,6 +14,7 @@
 
 extern crate khronos_api;
 
+use std::borrow::Cow;
 use std::collections::BTreeSet;
 use std::collections::HashMap;
 use std::fmt;
@@ -58,7 +59,7 @@ pub struct Enum {
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub struct Binding {
     pub ident: String,
-    pub ty: String,
+    pub ty: Cow<'static, str>,
 }
 
 #[derive(Clone, Debug, PartialEq, Eq)]
@@ -121,12 +122,12 @@ impl Registry {
 
     /// Returns a set of all the types used in the supplied registry. This is useful
     /// for working out what conversions are needed for the specific registry.
-    pub fn get_tys(&self) -> BTreeSet<String> {
+    pub fn get_tys(&self) -> BTreeSet<&str> {
         let mut tys = BTreeSet::new();
         for def in &self.cmds {
-            tys.insert(def.proto.ty.clone());
+            tys.insert(def.proto.ty.as_ref());
             for param in &def.params {
-                tys.insert(param.ty.clone());
+                tys.insert(param.ty.as_ref());
             }
         }
         tys
