+++ gl_generator/registry/parse.rs
@@ -14,6 +14,7 @@
 
 extern crate khronos_api;
 
+use std::borrow::Cow;
 use std::collections::hash_map::Entry;
 use std::collections::BTreeSet;
 use std::collections::HashSet;
@@ -119,7 +120,7 @@ fn trim_enum_prefix<'a>(ident: &'a str, api: Api) -> &'a str {
     }
 }
 
-fn trim_cmd_prefix<'a>(ident: &'a str, api: Api) -> &'a str {
+fn trim_cmd_prefix(ident: &str, api: Api) -> &str {
     match api {
         Api::Gl | Api::GlCore | Api::Gles1 | Api::Gles2 => trim_str(ident, "gl"),
         Api::Glx => trim_str(ident, "glX"),
@@ -516,7 +517,7 @@ trait Parse: Sized + Iterator<Item = ParseEvent> {
 
         Binding {
             ident: ident,
-            ty: ty,
+            ty: to_rust_ty(ty),
         }
     }
 }
@@ -615,6 +616,278 @@ impl FromXml for String {
     fn convert<P: Parse>(_: &mut P, a: &[Attribute]) -> String {
         get_attribute(a, "name").unwrap()
     }
+}
+
+/// Converts a C style type definition to the Rust equivalent
+pub fn to_rust_ty<T: AsRef<str>>(ty: T) -> Cow<'static, str> {
+    let ty = match ty.as_ref() {
+        // gl.xml types
+        "GLDEBUGPROC"               => "types::GLDEBUGPROC",
+        "GLDEBUGPROCAMD"            => "types::GLDEBUGPROCAMD",
+        "GLDEBUGPROCARB"            => "types::GLDEBUGPROCARB",
+        "GLDEBUGPROCKHR"            => "types::GLDEBUGPROCKHR",
+        "GLbitfield"                => "types::GLbitfield",
+        "GLboolean"                 => "types::GLboolean",
+        "GLbyte"                    => "types::GLbyte",
+        "GLclampd"                  => "types::GLclampd",
+        "GLclampf"                  => "types::GLclampf",
+        "GLclampx"                  => "types::GLclampx",
+        "GLdouble"                  => "types::GLdouble",
+        "GLeglImageOES"             => "types::GLeglImageOES",
+        "GLenum"                    => "types::GLenum",
+        "GLfixed"                   => "types::GLfixed",
+        "GLfloat"                   => "types::GLfloat",
+        "GLhalfNV"                  => "types::GLhalfNV",
+        "GLhandleARB"               => "types::GLhandleARB",
+        "GLint"                     => "types::GLint",
+        "GLint64"                   => "types::GLint64",
+        "GLint64EXT"                => "types::GLint64EXT",
+        "GLintptr"                  => "types::GLintptr",
+        "GLintptrARB"               => "types::GLintptrARB",
+        "GLshort"                   => "types::GLshort",
+        "GLsizei"                   => "types::GLsizei",
+        "GLsizeiptr"                => "types::GLsizeiptr",
+        "GLsizeiptrARB"             => "types::GLsizeiptrARB",
+        "GLsync"                    => "types::GLsync",
+        "GLubyte"                   => "types::GLubyte",
+        "GLuint"                    => "types::GLuint",
+        "GLuint64"                  => "types::GLuint64",
+        "GLuint64EXT"               => "types::GLuint64EXT",
+        "GLushort"                  => "types::GLushort",
+        "GLvdpauSurfaceNV"          => "types::GLvdpauSurfaceNV",
+        "void "                     => "()",
+        "GLboolean *"               => "*mut types::GLboolean",
+        "GLchar *"                  => "*mut types::GLchar",
+        "GLcharARB *"               => "*mut types::GLcharARB",
+        "GLdouble *"                => "*mut types::GLdouble",
+        "GLenum *"                  => "*mut types::GLenum",
+        "GLfixed *"                 => "*mut types::GLfixed",
+        "GLfloat *"                 => "*mut types::GLfloat",
+        "GLhandleARB *"             => "*mut types::GLhandleARB",
+        "GLint *"                   => "*mut types::GLint",
+        "GLint64 *"                 => "*mut types::GLint64",
+        "GLint64EXT *"              => "*mut types::GLint64EXT",
+        "GLsizei *"                 => "*mut types::GLsizei",
+        "GLubyte *"                 => "*mut types::GLubyte",
+        "GLuint *"                  => "*mut types::GLuint",
+        "GLuint [2]"                => "*mut [types::GLuint; 2]",
+        "GLuint64 *"                => "*mut types::GLuint64",
+        "GLuint64EXT *"             => "*mut types::GLuint64EXT",
+        "GLushort *"                => "*mut types::GLushort",
+        "GLvoid *"                  => "*mut types::GLvoid",
+        "GLvoid **"                 => "*const *mut types::GLvoid",
+        "void *"                    => "*mut __gl_imports::raw::c_void",
+        "void **"                   => "*const *mut __gl_imports::raw::c_void",
+        "const GLboolean *"         => "*const types::GLboolean",
+        "const GLbyte *"            => "*const types::GLbyte",
+        "const GLchar *"            => "*const types::GLchar",
+        "const GLcharARB *"         => "*const types::GLcharARB",
+        "const GLclampf *"          => "*const types::GLclampf",
+        "const GLdouble *"          => "*const types::GLdouble",
+        "const GLenum *"            => "*const types::GLenum",
+        "const GLfixed *"           => "*const types::GLfixed",
+        "const GLfloat"             => "types::GLfloat",
+        "const GLfloat *"           => "*const types::GLfloat",
+        "const GLhalfNV *"          => "*const types::GLhalfNV",
+        "const GLint *"             => "*const types::GLint",
+        "const GLint64 *"           => "*const types::GLint64",
+        "const GLint64EXT *"        => "*const types::GLint64EXT",
+        "const GLintptr *"          => "*const types::GLintptr",
+        "const GLshort *"           => "*const types::GLshort",
+        "const GLsizei *"           => "*const types::GLsizei",
+        "const GLsizeiptr *"        => "*const types::GLsizeiptr",
+        "const GLubyte *"           => "*const types::GLubyte",
+        "const GLuint *"            => "*const types::GLuint",
+        "const GLuint64 *"          => "*const types::GLuint64",
+        "const GLuint64EXT *"       => "*const types::GLuint64EXT",
+        "const GLushort *"          => "*const types::GLushort",
+        "const GLvdpauSurfaceNV *"  => "*const types::GLvdpauSurfaceNV",
+        "const GLvoid *"            => "*const types::GLvoid",
+        "const void *"              => "*const __gl_imports::raw::c_void",
+        "const void **"             => "*const *const __gl_imports::raw::c_void",
+        "const void *const*"        => "*const *const __gl_imports::raw::c_void",
+        "const GLboolean **"        => "*const *const types::GLboolean",
+        "const GLchar **"           => "*const *const types::GLchar",
+        "const GLcharARB **"        => "*const *const types::GLcharARB",
+        "const GLvoid **"           => "*const *const types::GLvoid",
+        "const GLchar *const*"      => "*const *const types::GLchar",
+        "const GLvoid *const*"      => "*const *const types::GLvoid",
+        "struct _cl_context *"      => "*const types::_cl_context",
+        "struct _cl_event *"        => "*const types::_cl_event",
+        "GLuint[2]"                 => "[Gluint; 2]",
+
+        // glx.xml types
+        "Bool"                      => "types::Bool",
+        "Colormap"                  => "types::Colormap",
+        "DMbuffer"                  => "types::DMbuffer",
+        "Font"                      => "types::Font",
+        "GLXContext"                => "types::GLXContext",
+        "GLXContextID"              => "types::GLXContextID",
+        "GLXDrawable"               => "types::GLXDrawable",
+        "GLXFBConfig"               => "types::GLXFBConfig",
+        "GLXFBConfigSGIX"           => "types::GLXFBConfigSGIX",
+        "GLXPbuffer"                => "types::GLXPbuffer",
+        "GLXPbufferSGIX"            => "types::GLXPbufferSGIX",
+        "GLXPixmap"                 => "types::GLXPixmap",
+        "GLXVideoCaptureDeviceNV"   => "types::GLXVideoCaptureDeviceNV",
+        "GLXVideoDeviceNV"          => "types::GLXVideoDeviceNV",
+        "GLXVideoSourceSGIX"        => "types::GLXVideoSourceSGIX",
+        "GLXWindow"                 => "types::GLXWindow",
+        // "GLboolean"                 => "types::GLboolean",
+        // "GLenum"                    => "types::GLenum",
+        // "GLint"                     => "types::GLint",
+        // "GLsizei"                   => "types::GLsizei",
+        // "GLuint"                    => "types::GLuint",
+        "Pixmap"                    => "types::Pixmap",
+        //"Status"                    => "types::Status",
+        //"VLNode"                    => "types::VLNode",
+        //"VLPath"                    => "types::VLPath",
+        //"VLServer"                  => "types::VLServer",
+        "Window"                    => "types::Window",
+        "__GLXextFuncPtr"           => "types::__GLXextFuncPtr",
+        "const GLXContext"          => "const types::GLXContext",
+        "float "                    => "__gl_imports::raw::c_float",
+        "int "                      => "__gl_imports::raw::c_int",
+        "int64_t"                   => "i64",
+        "unsigned int "             => "__gl_imports::raw::c_uint",
+        "unsigned long "            => "__gl_imports::raw::c_ulong",
+        // "void "                     => "()",
+        "DMparams *"                => "*mut types::DMparams",
+        "Display *"                 => "*mut types::Display",
+        "GLXFBConfig *"             => "*mut types::GLXFBConfig",
+        "GLXFBConfigSGIX *"         => "*mut types::GLXFBConfigSGIX",
+        "GLXHyperpipeConfigSGIX *"  => "*mut types::GLXHyperpipeConfigSGIX",
+        "GLXHyperpipeNetworkSGIX *" => "*mut types::GLXHyperpipeNetworkSGIX",
+        "GLXVideoCaptureDeviceNV *" => "*mut types::GLXVideoCaptureDeviceNV",
+        "GLXVideoDeviceNV *"        => "*mut types::GLXVideoDeviceNV",
+        // "GLuint *"                  => "*mut types::GLuint",
+        "XVisualInfo *"             => "*mut types::XVisualInfo",
+        // "const GLubyte *"           => "*GLubyte",
+        "const char *"              => "*const __gl_imports::raw::c_char",
+        "const int *"               => "*const __gl_imports::raw::c_int",
+        // "const void *"              => "*const __gl_imports::raw::c_void",
+        "int *"                     => "*mut __gl_imports::raw::c_int",
+        "int32_t *"                 => "*mut i32",
+        "int64_t *"                 => "*mut i64",
+        "long *"                    => "*mut __gl_imports::raw::c_long",
+        "unsigned int *"            => "*mut __gl_imports::raw::c_uint",
+        "unsigned long *"           => "*mut __gl_imports::raw::c_ulong",
+        // "void *"                    => "*mut __gl_imports::raw::c_void",
+
+        // wgl.xml types
+        "BOOL"                      => "types::BOOL",
+        "DWORD"                     => "types::DWORD",
+        "FLOAT"                     => "types::FLOAT",
+        // "GLbitfield"                => "types::GLbitfield",
+        // "GLboolean"                 => "types::GLboolean",
+        // "GLenum"                    => "types::GLenum",
+        // "GLfloat"                   => "types::GLfloat",
+        // "GLint"                     => "types::GLint",
+        // "GLsizei"                   => "types::GLsizei",
+        // "GLuint"                    => "types::GLuint",
+        // "GLushort"                  => "types::GLushort",
+        "HANDLE"                    => "types::HANDLE",
+        "HDC"                       => "types::HDC",
+        "HENHMETAFILE"              => "types::HENHMETAFILE",
+        "HGLRC"                     => "types::HGLRC",
+        "HGPUNV"                    => "types::HGPUNV",
+        "HPBUFFERARB"               => "types::HPBUFFERARB",
+        "HPBUFFEREXT"               => "types::HPBUFFEREXT",
+        "HPVIDEODEV"                => "types::HPVIDEODEV",
+        "HVIDEOINPUTDEVICENV"       => "types::HVIDEOINPUTDEVICENV",
+        "HVIDEOOUTPUTDEVICENV"      => "types::HVIDEOOUTPUTDEVICENV",
+        "INT"                       => "types::INT",
+        "INT64"                     => "types::INT64",
+        "LPCSTR"                    => "types::LPCSTR",
+        "LPGLYPHMETRICSFLOAT"       => "types::LPGLYPHMETRICSFLOAT",
+        "LPVOID"                    => "types::LPVOID",
+        "PGPU_DEVICE"               => "types::PGPU_DEVICE",
+        "PROC"                      => "types::PROC",
+        "UINT"                      => "types::UINT",
+        "VOID"                      => "types::VOID",
+        // "int "                      => "__gl_imports::raw::c_int",
+        // "unsigned int "             => "__gl_imports::raw::c_uint",
+        // "void "                     => "()",
+        "BOOL *"                    => "*mut types::BOOL",
+        "DWORD *"                   => "*mut types::DWORD",
+        "FLOAT *"                   => "*mut types::FLOAT",
+        // "GLuint *"                  => "*mut types::GLuint",
+        "HANDLE *"                  => "*mut types::HANDLE",
+        "HGPUNV *"                  => "*mut types::HGPUNV",
+        "HPVIDEODEV *"              => "*mut types::HPVIDEODEV",
+        "HVIDEOINPUTDEVICENV *"     => "*mut types::HVIDEOINPUTDEVICENV",
+        "HVIDEOOUTPUTDEVICENV *"    => "*mut types::HVIDEOOUTPUTDEVICENV",
+        "INT32 *"                   => "*mut types::INT32",
+        "INT64 *"                   => "*mut types::INT64",
+        "UINT *"                    => "*mut types::UINT",
+        "USHORT *"                  => "*mut types::USHORT",
+        "const COLORREF *"          => "*const types::COLORREF",
+        "const DWORD *"             => "*const types::DWORD",
+        "const FLOAT *"             => "*const types::FLOAT",
+        // "const GLushort *"          => "*const types::GLushort",
+        "const HANDLE *"            => "*const types::HANDLE",
+        "const HGPUNV *"            => "*const types::HGPUNV",
+        "const LAYERPLANEDESCRIPTOR *"  => "*const types::LAYERPLANEDESCRIPTOR",
+        "const LPVOID *"            => "*const types::LPVOID",
+        "const PIXELFORMATDESCRIPTOR *" => "*const types::IXELFORMATDESCRIPTOR",
+        "const USHORT *"            => "*const types::USHORT",
+        // "const char *"              => "*const __gl_imports::raw::c_char",
+        // "const int *"               => "*const __gl_imports::raw::c_int",
+        "float *"                   => "*mut __gl_imports::raw::c_float",
+        // "int *"                     => "*mut __gl_imports::raw::c_int",
+        // "unsigned long *"           => "*mut __gl_imports::raw::c_ulong",
+        // "void *"                    => "*mut __gl_imports::raw::c_void",
+
+        // elx.xml types
+        "khronos_utime_nanoseconds_t"   => "types::khronos_utime_nanoseconds_t",
+        "khronos_uint64_t"          => "types::khronos_uint64_t",
+        "khronos_ssize_t"           => "types::khronos_ssize_t",
+        "EGLNativeDisplayType"      => "types::EGLNativeDisplayType",
+        "EGLNativePixmapType"       => "types::EGLNativePixmapType",
+        "EGLNativeWindowType"       => "types::EGLNativeWindowType",
+        "EGLint"                    => "types::EGLint",
+        "EGLint *"                  => "*mut types::EGLint",
+        "const EGLint *"            => "*const types::EGLint",
+        "NativeDisplayType"         => "types::NativeDisplayType",
+        "NativePixmapType"          => "types::NativePixmapType",
+        "NativeWindowType"          => "types::NativeWindowType",
+        //"Bool"                      => "types::Bool",
+        "EGLBoolean"                => "types::EGLBoolean",
+        "EGLenum"                   => "types::EGLenum",
+        "EGLAttribKHR"              => "types::EGLAttribKHR",
+        "EGLAttrib"                 => "types::EGLAttrib",
+        "EGLAttrib *"               => "*mut types::EGLAttrib",
+        "const EGLAttrib *"         => "*const types::EGLAttrib",
+        "EGLConfig"                 => "types::EGLConfig",
+        "EGLConfig *"               => "*mut types::EGLConfig",
+        "EGLContext"                => "types::EGLContext",
+        "EGLDeviceEXT"              => "types::EGLDeviceEXT",
+        "EGLDisplay"                => "types::EGLDisplay",
+        "EGLSurface"                => "types::EGLSurface",
+        "EGLClientBuffer"               => "types::EGLClientBuffer",
+        "__eglMustCastToProperFunctionPointerType"  => "types::__eglMustCastToProperFunctionPointerType",
+        "EGLImageKHR"               => "types::EGLImageKHR",
+        "EGLImage"                  => "types::EGLImage",
+        "EGLSyncKHR"                => "types::EGLSyncKHR",
+        "EGLSync"                   => "types::EGLSync",
+        "EGLTimeKHR"                => "types::EGLTimeKHR",
+        "EGLTime"                   => "types::EGLTime",
+        "EGLSyncNV"                 => "types::EGLSyncNV",
+        "EGLTimeNV"                 => "types::EGLTimeNV",
+        "EGLuint64NV"               => "types::EGLuint64NV",
+        "EGLStreamKHR"              => "types::EGLStreamKHR",
+        "EGLuint64KHR"              => "types::EGLuint64KHR",
+        "EGLNativeFileDescriptorKHR"    => "types::EGLNativeFileDescriptorKHR",
+        "EGLsizeiANDROID"           => "types::EGLsizeiANDROID",
+        "EGLSetBlobFuncANDROID"     => "types::EGLSetBlobFuncANDROID",
+        "EGLGetBlobFuncANDROID"     => "types::EGLGetBlobFuncANDROID",
+        "EGLClientPixmapHI"         => "types::EGLClientPixmapHI",
+
+        // failure
+        _ => panic!("Type conversion not implemented for `{}`", ty.as_ref()),
+    };
+
+    Cow::Borrowed(ty)
 }
 
 #[cfg(test)]
